refactor: refactor and replace img to Image
