--- src/app/create-post/page.tsx
@@ -1,10 +1,0 @@
-// src/app/create-post/page.tsx
-'use client';
-
-import { PostCreationWizard } from '@/features/posts';
-import { useRouter } from 'next/navigation';
-
-export default function CreatePost(): React.ReactElement {
-  const router = useRouter();
-  return <PostCreationWizard exitSubscriberAction={() => router.back()} />;
-}
