+++ src/features/posts/hooks/create/useImageAspect.ts
@@ -1,8 +1,8 @@
 import { useCallback, useEffect } from 'react';
 import { useAppDispatch } from '@/shared/state/store';
-import { aspectRatios } from '@/shared/config/aspectRatios';
 import { PhotoSettings } from '../../lib/post.types';
 import { setPhotoSettings } from '../../model/postSlice';
+import { aspectRatios } from '../../lib/aspectRatios';
 
 type useImageAspectReturn = { getCropperAspect: () => number };
