+++ src/features/posts/ui/postCreate/AddPostSteps/FiltersStep/FiltersStep.tsx
@@ -22,7 +22,7 @@ export function FiltersStep({
     goPrev,
     hasNext,
     hasPrev,
-  } = useFiltersStep(onCloseAction);
+  } = useFiltersStep();
 
   return (
     <Modal
