+++ src/features/posts/ui/postView/ViewPost.tsx
@@ -24,8 +24,8 @@ export const ViewPost = ({ post, children }: Props): ReactNode => {
               src={photo}
               alt={'Post image'}
               key={index}
-              width={500}
-              height={500}
+              width={1000}
+              height={1000}
               unoptimized
             />
           );
